added development workflow guidelines.
Began writing development workflow guidelines for current and future developers to learn from. Small CSS tweak to display bullet list. Lots of work left for this page and those todos will be outlines in the issues. Images need to be added to the guide for newbies. Details need to be fleshed out and the importance of GitHub needs to be emphasized.